requirements.txt
altair==6.0.0
annotated-types==0.7.0   
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3   
blinker==1.9.0
cachetools==6.2.4        
certifi==2025.11.12      
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3     
distro==1.9.0
docx2pdf==0.1.8
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.14.0
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
streamlit==1.52.2
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
watchdog==6.0.0
xxhash==3.6.0
zstandard==0.25.0